requirements.txt
altgraph==0.17.4
attrs==25.2.0
blinker==1.9.0
bottle==0.13.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
clr_loader==0.2.7.post0
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
narwhals==1.30.0
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pillow==11.1.0
plotly==6.0.0
protobuf==5.29.3
proxy_tools==0.1.0
pyarrow==19.0.1
pycparser==2.22
pydeck==0.9.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
python-dateutil==2.9.0.post0
pythonnet==3.0.5
pytz==2025.1
pywebview==5.4
pywin32-ctypes==0.2.3
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.38
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
Werkzeug==3.1.3
streamlit==1.18.1